+++ lib/src/components/components.dart
@@ -2,6 +2,13 @@
 // Use of this source code is governed by a MIT-style license
 // that can be found in the LICENSE file.
 
+export 'common_components.dart';
 export 'day_view_components.dart';
+export 'event_scroll_notifier.dart';
+export 'headers/calendar_page_header.dart';
+export 'headers/day_page_header.dart';
+export 'headers/month_page_header.dart';
+export 'headers/week_page_header.dart';
 export 'month_view_components.dart';
+export 'safe_area_wrapper.dart';
 export 'week_view_components.dart';